requirements.txt
anyio==3.6.2
astroid==2.12.13
attrs==22.1.0
bcrypt==4.0.1
black==22.12.0
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
dill==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.87.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
orjson==3.8.2
packaging==21.3
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pylint==2.15.9
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.3
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
ujson==5.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
